make document block component an island
